% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_reference_bases.R
\name{extract_reference_bases}
\alias{extract_reference_bases}
\title{Extract Reference Bases for Edit Sites}
\usage{
extract_reference_bases(genomic_sequence, chromosome, start, stop)
}
\arguments{
\item{genomic_sequence}{A \code{BSgenome} object containing the genomic sequence from which reference bases will be extracted.}

\item{chromosome}{Chromosome of edit site}

\item{start}{Start position for edit site. Relative to + strand of genomic sequence.}

\item{stop}{Stop position edit site. Relative to + strand of genomic sequence.}
}
\value{
table with a column with the reference base and its genomic location
}
\description{
This function extracts the genomic sequence for a defined chromosome with start and stop coordinates. Does not consider strand, assumes coordinates are in a plus strand orientation relative to the provided genomic sequence.
}
