% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bam_alignments.R
\name{filter_bam_alignments}
\alias{filter_bam_alignments}
\title{Filter BAM Alignments}
\usage{
filter_bam_alignments(
  df_bam,
  specimen,
  chromosome,
  edit_site_position,
  edit_site_strand,
  allowed_aln_start_cut_site_PAM_distal = 3,
  allowed_aln_start_cut_site_PAM_proximal = 3
)
}
\arguments{
\item{df_bam}{A data frame containing BAM alignments (must contain \code{cigar} and \code{width} columns).}

\item{specimen}{The specimen ID.}

\item{chromosome}{The chromosomal location of the edit site.}

\item{edit_site_position}{The position of the edit site.}

\item{edit_site_strand}{The strand of the edit site ('+' or '-').}

\item{allowed_aln_start_cut_site_PAM_distal}{The allowed alignment start with respect to the cut site, PAM distal. Default is 3.}

\item{allowed_aln_start_cut_site_PAM_proximal}{The allowed alignment start with respect to the cut site, PAM proximal Default is 3.}
}
\value{
A filtered data frame of BAM alignments.
}
\description{
Takes as input a data frame of BAM alignments and filters based on edit site position and strand.
}
\details{
This function filters BAM alignments based on the specified edit site position and strand.

The function uses different filtering criteria based on the strand of the edit site:
\itemize{
\item For the '+' strand, alignments are filtered with flag '83', indicating the cut site is to the left of the PAM site.
\item For the '-' strand, alignments are filtered with flag '99', indicating the cut site is to the right of the PAM site.
}

Note: \code{pos} represents the 1-based start position of the alignment on the genome.
We use the genomic alignment \code{width} (not \code{qwidth}) to calculate the exact genomic end position (\code{pos + width - 1}).

Outputted columns \code{edit_site_min} and \code{edit_site_max} are always relative to the plus strand of the genome.
}
