% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/protospacer_coords_to_absolute_start_stop_coords.R
\name{protospacer_coords_to_absolute_start_stop_coords}
\alias{protospacer_coords_to_absolute_start_stop_coords}
\title{Convert Protospacer and PAM Coordinates to Absolute Start and Stop Coordinates}
\usage{
protospacer_coords_to_absolute_start_stop_coords(protospacer_coords, strand)
}
\arguments{
\item{protospacer_coords}{A list containing \code{protospacer_end} and \code{PAM_start} coordinates. Can be generated by \code{cut_site_protospacer_coordinates()}.}

\item{strand}{Character. The strand orientation, either \code{'+'} or \code{'-'}.}
}
\value{
A named list containing:
\describe{
\item{start}{The absolute minimum genomic coordinate of the target site.}
\item{stop}{The absolute maximum genomic coordinate of the target site.}
}
}
\description{
This function takes calculated protospacer/PAM boundaries and converts them
into absolute, 1-based genomic start (minimum) and stop (maximum) coordinates
spanning the entire target site (including the full PAM).
}
