% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_cut_site_direction.R
\name{standardize_cut_site_direction}
\alias{standardize_cut_site_direction}
\title{Standardize Cut Site Orientation}
\usage{
standardize_cut_site_direction(strand, relative_zero, genomic_positions, bases)
}
\arguments{
\item{strand}{Character. The strand orientation of the cut site (\code{'+'} or \code{'-'}).}

\item{relative_zero}{Numeric. The genomic coordinate representing the zero relative position (position zero).}

\item{genomic_positions}{Numeric vector. The genomic positions of each base to standardize.}

\item{bases}{Character vector. The base calls corresponding to each genomic position.}
}
\value{
A \code{data.frame} with columns:
\describe{
\item{genomic_position}{Original genomic position of each base.}
\item{base}{Base at the given position (reverse-complemented if strand is \code{'-'}).}
\item{relative_position}{Position relative to the cut site. Is a factor.}
}
}
\description{
Adjusts genomic positions and base identities relative to a specified genomic position
(relative zero) according to the provided strand orientation. All relative positions read in the
5'->3' direction.
For the positive strand, relative positions are computed as \code{relative_zero - genomic_position}. Genomic positions upstream
of the relative zero will have positive relative position values and downstream will be negative.
For the negative strand, relative positions are computed as
\code{genomic_position - relative_zero}, and \verb{bases are complemented}. Genomic positions downstream of the relative zero
will have positive relative position values and upstream will be negative.
}
\examples{
\dontrun{
genomic_positions <- 100:105
bases <- c("A", "T", "G", "C", "C", "A")
standardize_cut_site_direction(
  strand = '+',
  relative_zero = 103,
  genomic_positions = genomic_positions,
  bases = bases
)

standardize_cut_site_direction(
  strand = '-',
  relative_zero = 103,
  genomic_positions = genomic_positions,
  bases = bases
)
}

}
