% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_site_to_protospacer_coordinates.R
\name{cut_site_to_protospacer_coordinates}
\alias{cut_site_to_protospacer_coordinates}
\title{Convert a cut site position to protospacer and PAM coordinates}
\usage{
cut_site_to_protospacer_coordinates(
  cut_site,
  strand,
  end_distance_from_cut_site = 20,
  expected_cut_distance_from_pam = 3,
  pam_length = 3
)
}
\arguments{
\item{cut_site}{Numeric. The cut site position (1-based).}

\item{strand}{Character. The strand orientation, either \code{'+'} or \code{'-'}.}

\item{end_distance_from_cut_site}{Numeric. Distance from the cut site to the end of the protospacer. Default is 20.}

\item{expected_cut_distance_from_pam}{Numeric. Distance between the cut site and the end of expected PAM site. Default is 3.}

\item{pam_length}{Numeric. Length of the PAM sequence. Default is 3.}
}
\value{
A named list with elements:
\describe{
\item{PAM_start}{The calculated start position of the PAM.}
\item{protospacer_start}{The calculated start position of the protospacer.}
\item{protospacer_end}{The calculated end position of the protospacer.}
}
}
\description{
Nomenclature visual: 5'->protospacer_end--------cut_site---protospacer_startPAM_end-PAM_start->3'
}
\details{
Given a genomic cut site and strand orientation, this function calculates the
coordinates of the protospacer start, protospacer end, and PAM start.
}
\examples{
\dontrun{
cut_site_to_protospacer_coordinates(100, '+')
cut_site_to_protospacer_coordinates(100, '-', end_distance_from_cut_site = 25)
}

}
